requirements.txt
aiomysql==0.0.20
argon2-cffi==19.2.0
cffi==1.14.0
cryptography==2.9
pycparser==2.20
PyMySQL==0.9.2
six==1.14.0
tornado==6.0.4
